Move trait into Concerns sub-namespace
