+++ packages/framework/src/Framework/Features/Publications/Models/PublicationFieldValues/TextField.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Features\Publications\Models\PublicationFieldValues;
 
+use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Concerns\CanonicableTrait;
 use Hyde\Framework\Features\Publications\Models\PublicationFieldValues\Contracts\Canonicable;
 use Hyde\Framework\Features\Publications\PublicationFieldTypes;
+
 use function trim;
 
 final class TextField extends PublicationFieldValue implements Canonicable
